Make repertoire the home page
