+++ app/page.tsx
@@ -1,122 +1,268 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { findMatches, SingerEntry } from "@/lib/matching";
+import type { Confidence, Part, Voicing } from "@/lib/matching";
+import {
+  addRepertoireItem,
+  deleteRepertoireItem,
+  getMyRepertoire,
+  type RepertoireRow,
+} from "@/lib/repertoireStore";
+import { getCurrentUser } from "@/lib/profileStore";
 
-type RepertoireItem = {
-  id: string;
-  songTitle: string;
-  voicing: "TTBB" | "SATB" | "SSAA";
-  arrangerName?: string;
-  partsKnown: any[];
-  confidence: any;
+const voicings: Voicing[] = ["TTBB", "SATB", "SSAA"];
+
+const partsByVoicing: Record<Voicing, Part[]> = {
+  TTBB: ["Tenor", "Lead", "Baritone", "Bass"],
+  SATB: ["Soprano", "Alto", "Tenor", "Bass"],
+  SSAA: ["Soprano 1", "Soprano 2", "Alto 1", "Alto 2"],
 };
 
+const confidenceLevels: Confidence[] = [
+  "Performance ready",
+  "Solid",
+  "Needs review",
+  "Rusty",
+  "Learning",
+];
+
 export default function Home() {
-  const [entries, setEntries] = useState<SingerEntry[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [items, setItems] = useState<RepertoireRow[]>([]);
+  const [songTitle, setSongTitle] = useState("");
+  const [voicing, setVoicing] = useState<Voicing>("TTBB");
+  const [arrangerName, setArrangerName] = useState("");
+  const [partsKnown, setPartsKnown] = useState<Part[]>([]);
+  const [confidence, setConfidence] = useState<Confidence>("Solid");
+  const [message, setMessage] = useState("");
+
+  async function loadRepertoire() {
+    const user = await getCurrentUser();
+
+    if (!user) {
+      window.location.href = "/login";
+      return;
+    }
+
+    const data = await getMyRepertoire();
+    setItems(data);
+  }
 
   useEffect(() => {
-    try {
-      const saved = window.localStorage.getItem("what-can-we-sing-repertoire");
-      if (!saved) return;
-
-      const repertoire: RepertoireItem[] = JSON.parse(saved);
-
-      const converted: SingerEntry[] = repertoire.map((item) => ({
-        userId: "me",
-        displayName: "You",
-        songTitle: item.songTitle,
-        voicing: item.voicing,
-        arrangerName: item.arrangerName ?? null,
-        partsKnown: item.partsKnown,
-        confidence: item.confidence,
-      }));
-
-      setEntries(converted);
-    } catch {
-      console.warn("Failed to load repertoire");
+    async function load() {
+      try {
+        await loadRepertoire();
+      } catch (err) {
+        console.error(err);
+        setMessage("Could not load repertoire.");
+      } finally {
+        setLoading(false);
+      }
     }
+
+    load();
   }, []);
 
-  const matches = findMatches(entries);
+  function togglePart(part: Part) {
+    setPartsKnown((current) =>
+      current.includes(part)
+        ? current.filter((p) => p !== part)
+        : [...current, part]
+    );
+  }
+
+  async function addItem() {
+    if (!songTitle.trim() || partsKnown.length === 0) return;
+
+    try {
+      await addRepertoireItem({
+        songTitle: songTitle.trim(),
+        voicing,
+        arrangerName: arrangerName.trim() || undefined,
+        partsKnown,
+        confidence,
+      });
+
+      setSongTitle("");
+      setArrangerName("");
+      setPartsKnown([]);
+      setConfidence("Solid");
+      setMessage("Song added.");
+
+      await loadRepertoire();
+    } catch (err) {
+      console.error(err);
+      setMessage("Could not add song.");
+    }
+  }
+
+  async function deleteItem(id: string) {
+    try {
+      await deleteRepertoireItem(id);
+      setMessage("Song deleted.");
+      await loadRepertoire();
+    } catch (err) {
+      console.error(err);
+      setMessage("Could not delete song.");
+    }
+  }
+
+  if (loading) {
+    return (
+      <main className="flex min-h-screen items-center justify-center bg-slate-950 text-white">
+        Loading repertoire...
+      </main>
+    );
+  }
 
   return (
     <main className="min-h-screen bg-slate-950 px-6 py-10 text-white">
-      <div className="mx-auto max-w-4xl">
-        <div className="flex gap-4">
-          <a href="/repertoire" className="text-sm text-cyan-300 hover:text-cyan-200">
-            → Manage my repertoire
-          </a>
+      <div className="mx-auto max-w-5xl">
+        <div className="flex flex-wrap gap-4">
           <a href="/session" className="text-sm text-cyan-300 hover:text-cyan-200">
             → Start session
           </a>
           <a href="/settings" className="text-sm text-cyan-300 hover:text-cyan-200">
             → Settings
           </a>
-          <a href="/login" className="text-sm text-cyan-300 hover:text-cyan-200">
-            → Log in
-          </a>
         </div>
 
-        <p className="mt-4 text-sm font-semibold uppercase tracking-wide text-cyan-300">
-          What Can We Sing
+        <h1 className="mt-4 text-4xl font-bold tracking-tight">My Repertoire</h1>
+        <p className="mt-2 text-slate-300">
+          Add songs you know, the parts you can sing, and how confident you are.
         </p>
 
-        <h1 className="mt-3 text-4xl font-bold tracking-tight">
-          Your songs (solo view for now)
-        </h1>
+        {message && <p className="mt-4 text-sm text-slate-300">{message}</p>}
 
-        {entries.length === 0 && (
-          <p className="mt-6 text-slate-300">
-            You haven’t added any songs yet. Go to “Manage my repertoire.”
-          </p>
-        )}
+        <section className="mt-8 rounded-2xl border border-white/10 bg-white/10 p-5 shadow-lg">
+          <h2 className="text-2xl font-semibold">Add a song</h2>
 
-        <div className="mt-8 space-y-4">
-          {matches.map((match) => (
-            <section
-              key={`${match.songTitle}-${match.voicing}`}
-              className="rounded-2xl border border-white/10 bg-white/10 p-5 shadow-lg"
-            >
-              <div className="flex items-start justify-between gap-4">
+          <div className="mt-5 grid gap-4 md:grid-cols-2">
+            <label className="block">
+              <span className="text-sm font-medium text-slate-300">Song title</span>
+              <input
+                value={songTitle}
+                onChange={(e) => setSongTitle(e.target.value)}
+                placeholder="Hello Mary Lou"
+                className="mt-1 w-full rounded-xl border border-white/10 bg-slate-900 px-4 py-3 text-white outline-none ring-cyan-300 focus:ring-2"
+              />
+            </label>
+
+            <label className="block">
+              <span className="text-sm font-medium text-slate-300">
+                Arranger optional
+              </span>
+              <input
+                value={arrangerName}
+                onChange={(e) => setArrangerName(e.target.value)}
+                placeholder="Unknown is okay"
+                className="mt-1 w-full rounded-xl border border-white/10 bg-slate-900 px-4 py-3 text-white outline-none ring-cyan-300 focus:ring-2"
+              />
+            </label>
+
+            <label className="block">
+              <span className="text-sm font-medium text-slate-300">Voicing</span>
+              <select
+                value={voicing}
+                onChange={(e) => {
+                  setVoicing(e.target.value as Voicing);
+                  setPartsKnown([]);
+                }}
+                className="mt-1 w-full rounded-xl border border-white/10 bg-slate-900 px-4 py-3 text-white outline-none ring-cyan-300 focus:ring-2"
+              >
+                {voicings.map((v) => (
+                  <option key={v}>{v}</option>
+                ))}
+              </select>
+            </label>
+
+            <label className="block">
+              <span className="text-sm font-medium text-slate-300">Confidence</span>
+              <select
+                value={confidence}
+                onChange={(e) => setConfidence(e.target.value as Confidence)}
+                className="mt-1 w-full rounded-xl border border-white/10 bg-slate-900 px-4 py-3 text-white outline-none ring-cyan-300 focus:ring-2"
+              >
+                {confidenceLevels.map((level) => (
+                  <option key={level} value={level}>
+                    {level}
+                  </option>
+                ))}
+              </select>
+            </label>
+          </div>
+
+          <div className="mt-5">
+            <p className="text-sm font-medium text-slate-300">Parts you know</p>
+            <div className="mt-2 flex flex-wrap gap-2">
+              {partsByVoicing[voicing].map((part) => (
+                <button
+                  key={part}
+                  type="button"
+                  onClick={() => togglePart(part)}
+                  className={`rounded-full px-4 py-2 text-sm font-semibold ${
+                    partsKnown.includes(part)
+                      ? "bg-cyan-300 text-slate-950"
+                      : "bg-slate-800 text-slate-200 hover:bg-slate-700"
+                  }`}
+                >
+                  {part}
+                </button>
+              ))}
+            </div>
+          </div>
+
+          <button
+            onClick={addItem}
+            className="mt-6 rounded-xl bg-cyan-300 px-5 py-3 font-semibold text-slate-950 hover:bg-cyan-200 disabled:opacity-40"
+            disabled={!songTitle.trim() || partsKnown.length === 0}
+          >
+            Add to repertoire
+          </button>
+        </section>
+
+        <section className="mt-8">
+          <h2 className="text-2xl font-semibold">Songs I know</h2>
+
+          <div className="mt-4 space-y-3">
+            {items.length === 0 && (
+              <p className="rounded-2xl border border-white/10 bg-white/5 p-5 text-slate-300">
+                No songs yet. Add your first one above.
+              </p>
+            )}
+
+            {items.map((item) => (
+              <div
+                key={item.id}
+                className="flex flex-col justify-between gap-4 rounded-2xl border border-white/10 bg-white/10 p-5 shadow-lg md:flex-row md:items-center"
+              >
                 <div>
-                  <h2 className="text-2xl font-semibold">
-                    {match.songTitle} — {match.voicing}
-                  </h2>
+                  <h3 className="text-xl font-semibold">
+                    {item.song_title} — {item.voicing}
+                  </h3>
                   <p className="mt-1 text-sm text-slate-300">
-                    {match.category === "ready" && "You can cover all parts"}
-                    {match.category === "possible" && "Possible match"}
-                    {match.category === "one_part_missing" &&
-                      `Missing: ${match.missingParts.join(", ")}`}
+                    {item.arranger_name
+                      ? `Arr. ${item.arranger_name}`
+                      : "Arranger unknown"}
+                  </p>
+                  <p className="mt-2 text-sm text-slate-200">
+                    Parts: {item.parts_known.join(", ")}
+                  </p>
+                  <p className="text-sm text-slate-300">
+                    Confidence: {item.confidence}
                   </p>
                 </div>
 
-                <span className="rounded-full bg-cyan-300 px-3 py-1 text-sm font-semibold text-slate-950">
-                  {match.category}
-                </span>
+                <button
+                  onClick={() => deleteItem(item.id)}
+                  className="rounded-xl bg-rose-400/10 px-4 py-2 text-sm font-semibold text-rose-200 hover:bg-rose-400/20"
+                >
+                  Delete
+                </button>
               </div>
-
-              <div className="mt-4 grid gap-3 sm:grid-cols-2">
-                {Object.entries(match.assignments).map(([part, singers]) => (
-                  <div key={part} className="rounded-xl bg-slate-900/70 p-3">
-                    <p className="font-semibold">{part}</p>
-                    <p className="text-sm text-slate-300">
-                      {singers.map((s) => s.displayName).join(", ")}
-                    </p>
-                  </div>
-                ))}
-              </div>
-
-              {match.warnings.length > 0 && (
-                <div className="mt-4 rounded-xl bg-amber-300/10 p-3 text-sm text-amber-200">
-                  {match.warnings.map((warning) => (
-                    <p key={warning}>⚠ {warning}</p>
-                  ))}
-                </div>
-              )}
-            </section>
-          ))}
-        </div>
+            ))}
+          </div>
+        </section>
       </div>
     </main>
   );
